+++ tests/src/python/CMakeLists.txt
@@ -257,7 +257,8 @@ ADD_PYTHON_TEST(PyQgsPointCloudLayerProfileGenerator test_qgspointcloudlayerprof
 ADD_PYTHON_TEST(PyQgsPointCloudRgbRenderer test_qgspointcloudrgbrenderer.py)
 ADD_PYTHON_TEST(PyQgsPointClusterRenderer test_qgspointclusterrenderer.py)
 ADD_PYTHON_TEST(PyQgsPointDisplacementRenderer test_qgspointdisplacementrenderer.py)
-ADD_PYTHON_TEST(PyQgsProcessExecutable test_qgsprocessexecutable.py)
+ADD_PYTHON_TEST(PyQgsProcessExecutablePt1 test_qgsprocessexecutable_pt1.py)
+ADD_PYTHON_TEST(PyQgsProcessExecutablePt2 test_qgsprocessexecutable_pt2.py)
 ADD_PYTHON_TEST(PyQgsProcessingRecentAlgorithmLog test_qgsprocessingrecentalgorithmslog.py)
 ADD_PYTHON_TEST(PyQgsProcessingInPlace test_qgsprocessinginplace.py)
 ADD_PYTHON_TEST(PyQgsProcessingAlgRunner test_qgsprocessingalgrunner.py)
